// app/api/agent/types.ts
export type ComputerCallOutput = {
  type: "computer_call_output";
  call_id: string;
  output: { type: "input_image"; image_url: string };
  acknowledged_safety_checks: SafetyCheck[];
  current_url?: string;
};
// ...
export type Item = Message | FunctionToolCall | ComputerToolCall | Reasoning;

export type Message = {
  id: string;
  type: "message";
  role: "user" | "assistant" | "developer" | "system";
  content: Content[];
};

export type Reasoning = {
  id: string;
  type: "reasoning";
  summary: {
    type: "summary_text";
    text: string;
  }[];
};

export type FunctionToolCall = {
  type: "function_call";
  id: string;
  call_id: string;
  name: string;
  arguments: string;
  output: Content[] | null;
};

export type ComputerAction =
  | Click
  | DoubleClick
  | Drag
  | Screenshot
  | KeyPress
  | Move
  | Scroll
  | Type
  | Wait;

export type ComputerToolCall = {
  type: "computer_call";
  id: string;
  call_id: string;
  action: ComputerAction;
  pending_safety_checks: SafetyCheck[];
};

export type Click = {
  type: "click";
  button: "left" | "right" | "wheel" | "back" | "forward";
  x: number;
  y: number;
};

export type DoubleClick = {
  type: "double_click";
  x: number;
  y: number;
};

export type Scroll = {
  type: "scroll";
  x: number;
  y: number;
  scroll_x: number;
  scroll_y: number;
};

export type Type = {
  type: "type";
  text: string;
};

export type Wait = {
  type: "wait";
};

export type KeyPress = {
  type: "keypress";
  keys: string[];
};

export type Drag = {
  type: "drag";
  path: {
    x: number;
    y: number;
  }[];
};

export type Screenshot = {
  type: "screenshot";
};

export type Move = {
  type: "move";
  x: number;
  y: number;
};

export type SafetyCheck = {
  id: string;
  code: string;
  message: string;
};

export type InputContent = InputText | InputImage | InputFile;

export type OutputContent = OutputText | Refusal;

export type Content = InputContent | OutputContent | Reasoning;